add mask as channel in preprocessing (#219)
* fixed bug from removing flatfield

* cleaned removal of untracked array store

* removed untracked array metadata writing

* add channel metadata to zattrs

* updated doc for preprocessing

* zarr structure modified by suggestion

* added well format with only FOVs

* transfer io_utils to meta_utils

* fixed accidental edits

* formatted files

* moved add_channel function to mp_utils

* linting preprocessing readme

* fixed zarr structure description
